+++ packages/expo-router/src/link/useLinkHooks.ts
@@ -1,6 +1,6 @@
 // Fork of @react-navigation/native Link.tsx with `href` and `replace` support added and
 // `to` / `action` support removed.
-import { useMemo, MouseEvent } from 'react';
+import { useMemo, MouseEvent, ReactNode } from 'react';
 import { TextProps, GestureResponderEvent, Platform } from 'react-native';
 
 import { Href } from '../types';
@@ -210,6 +210,13 @@ export interface LinkProps extends Omit<TextProps, 'href'>, WebAnchorProps {
    * Replaces the initial screen with the current route.
    */
   withAnchor?: boolean;
+
+  /**
+   * Preview the route inside a context menu before navigating.
+   */
+  preview?: boolean;
+
+  previewItems?: ReactNode[];
 }
 
 // Mutate the style prop to add the className on web.